requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
certifi==2017.7.27.1      # via requests
chardet==3.0.4            # via requests
click==6.7                # via pip-tools
codecov==2.0.9
coverage==4.4.1
django-cors-headers==2.1.0
django-filter==1.0.4
django==1.11.5
djangorestframework==3.6.4
first==2.0.1              # via pip-tools
idna==2.6                 # via requests
markdown==2.6.9
pip-tools==1.9.0
pytz==2017.2              # via django
requests==2.18.4          # via codecov
six==1.11.0               # via pip-tools
urllib3==1.22             # via requests
